requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
audioop-lts==0.2.2
audioread==3.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
coqpit-config==0.2.3
coqui-tts==0.27.5
coqui-tts-trainer==0.3.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
docopt==0.6.2
dotenv==0.9.9
einops==0.8.2
executing==2.2.1
filelock==3.20.3
filetype==1.2.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
google-auth==2.47.0
google-genai==1.60.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.4
idna==3.11
inflect==7.5.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
ko-speech-tools==0.1.0
langchain-core==1.2.7
langchain-google-genai==4.2.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.4
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
Markdown==3.10.1
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
monotonic-alignment-search==0.2.1
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
num2words==0.5.14
numba==0.63.1
numpy==2.3.5
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
pooch==1.8.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.2
shellingham==1.5.4
simpleaudio==1.0.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
torchcodec==0.10.0
tqdm==4.67.1
traitlets==5.14.3
transformers==5.0.0
typeguard==4.4.4
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
wcwidth==0.3.2
websockets==15.0.1
Werkzeug==3.1.5
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
